Bug fixing, all the tests up to 9 pass.

+++ lib/include/mmu/mmu.h
@@ -41,7 +41,7 @@ class MMU
 
     auto write(uint16_t address, uint8_t value) -> void;
 
-    auto write16(uint16_t address, uint8_t value) -> void;
+    auto write16(uint16_t address, uint16_t value) -> void;
 
     auto ie_register() -> uint8_t;
